requirements.txt
pycryptodomex
pyzipper
